+++ starter/nuxt-auth/components/auth/AuthLayout.vue
@@ -1,0 +1,6 @@
+<template>
+  <AuthShell>
+    <AuthSidebar />
+    <slot />
+  </AuthShell>
+</template>