+++ starter/nuxt-auth/pages/auth/login.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+definePageMeta({
+  auth: false,
+  middleware: 'guest',
+});
+</script>
+
+<template>
+  <AuthLayout>
+    <AuthLogin />
+  </AuthLayout>
+</template>